Added routing, react-router-dom v6
