+++ src/App.js
@@ -1,36 +1,15 @@
-import Categories from "./components/Categories/Categories.component";
+import {Routes, Route} from "react-router-dom";
+import Navigation from "./routes/navigation/Navigation.component";
+import Home  from "./routes/home/Home.component";
 
 const App = () => {
-  const categories = [
-    {
-      "id": 1,
-      "title": "hats",
-      "imageUrl": "https://i.ibb.co/cvpntL1/hats.png"
-    },
-    {
-      "id": 2,
-      "title": "jackets",
-      "imageUrl": "https://i.ibb.co/px2tCc3/jackets.png"
-    },
-    {
-      "id": 3,
-      "title": "sneakers",
-      "imageUrl": "https://i.ibb.co/0jqHpnp/sneakers.png"
-    },
-    {
-      "id": 4,
-      "title": "womens",
-      "imageUrl": "https://i.ibb.co/GCCdy8t/womens.png"
-    },
-    {
-      "id": 5,
-      "title": "mens",
-      "imageUrl": "https://i.ibb.co/R70vBrQ/men.png"
-    }
-  ];
-
   return (
-    <Categories categories={categories}  />
+    <Routes>
+      <Route path="/" element={<Navigation/>} >
+      <Route index element={<Home />} />  
+      <Route path="shop"/>
+      </Route>
+    </Routes>
   );
 };
