+++ src/index.scss
@@ -11,3 +11,12 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+a {
+text-decoration: none;
+color: black;
+}
+
+*{
+  box-sizing: border-box;
+}
