+++ src/routes/navigation/Navigation.component.jsx
@@ -1,0 +1,23 @@
+import { Fragment } from "react";
+import { Outlet, Link } from "react-router-dom";
+import {ReactComponent as Logo} from "../../assets/crown.svg";
+import "./Navigation.styles.scss";
+
+const Navigation = () => {
+  return (
+    <Fragment>
+      <div className="navigation">
+        <Link className="logo-container" to="/">
+          <Logo className="logo"/>
+        </Link>
+        <div className="nav-links-container">
+          <Link className="nav-link" to="/shop">SHOP</Link>
+        </div>
+      </div>
+
+      <Outlet />
+    </Fragment>
+  );
+};
+
+export default Navigation;
